requirements.txt
streamlit==1.31.1
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
langchain-community==0.0.13
sentence-transformers==2.2.2
PyMuPDF==1.23.8
torch==2.1.0
transformers==4.36.2
huggingface-hub==0.20.3
pydub==0.25.1
audiorecorder==0.1
Pillow==10.2.0
requests==2.31.0
